build: bundle the Apple transcription sidecar
